requirements/base.txt
celery>=4.2.1
channels-redis>=2.4.0
channels>=2.2.0
cryptography>=2.9.2
Django>=3.0.8
django-cors-headers>=2.2.0
django-configurations>=2.0
django-cleanup>=4.0.0
django-filter>=2.0.0
djangorestframework>=3.11.0
djangorestframework-jwt>=1.11.0
djangorestframework-xml>=1.3.0
drf-flex-fields>=0.5.0
gunicorn>=20.0.4
redis>=3.2.0